added IndexReader getmanifest test, improved PackageList tests.
