+++ src/Tetrifact.Tests/PackageList/Base.cs
@@ -9,6 +9,8 @@ namespace Tetrifact.Tests.PackageList
     {
         protected IPackageList PackageList { get; private set; }
 
+        protected TestLogger<IPackageList> PackageListLogger { get; private set; }
+
         private IMemoryCache _memoryCache;
 
         public Base()
@@ -17,8 +19,8 @@ namespace Tetrifact.Tests.PackageList
             services.AddMemoryCache();
             ServiceProvider serviceProvider = services.BuildServiceProvider();
             _memoryCache = serviceProvider.GetService<IMemoryCache>();
-
-            this.PackageList = new Core.PackageList(_memoryCache, Settings, new TestLogger<IPackageList>());
+            this.PackageListLogger = new TestLogger<IPackageList>();
+            this.PackageList = new Core.PackageList(_memoryCache, Settings, this.PackageListLogger);
         }
 
         public void Dispose()
